+++ src/app.ts
@@ -2,117 +2,252 @@
  * Created by Caleydo Team on 31.08.2016.
  */
 
+import {listAll, IDType} from 'phovea_core/src/idtype';
+import {select} from 'd3';
+import ColumnManager, {IMotherTableType, AnyColumn} from './column/ColumnManager';
+import SupportView from './SupportView';
+import {Range1D} from 'phovea_core/src/range';
+import {EOrientation} from './column/AColumn';
+import MatrixFilter from './filter/MatrixFilter';
 import * as d3 from 'd3';
-import {IDataType} from 'phovea_core/src/datatype';
-import {list as listData, convertTableToVectors} from 'phovea_core/src/data';
-import {choose} from 'phovea_ui/src/dialogs';
-import {randomId} from 'phovea_core/src/index';
-import VisManager from './VisManager';
-import FilterManager from './FilterManager';
-import RangeManager from './RangeManager';
-import Block from './Block';
-import ConnectionLines from './ConnectionLines';
-
+import MatrixColumn from './column/MatrixColumn';
+import FilterManager from './filter/FilterManager';
+import {AVectorColumn} from './column/AVectorColumn';
+import {IAnyVector} from 'phovea_core/src/vector';
 /**
  * The main class for the App app
  */
+
 export default class App {
 
-  private readonly $node;
-  public static blockList = new Map();
-  public static visNode;
-  public static filterNode;
+  private readonly node: HTMLElement;
 
-  private visManager: VisManager;
-  private filterManager: FilterManager;
-  private rangeManager: RangeManager;
-  private connectionLines: ConnectionLines;
 
-  constructor(parent: Element) {
-    this.$node = d3.select(parent);
-    this.$node.select('main').append('div').classed('visManager', true);
-    App.visNode = d3.select('.visManager');
-    App.filterNode = d3.select('#filterView');
-    this.visManager = new VisManager();
-    this.rangeManager = new RangeManager(this.visManager);
-    this.filterManager = new FilterManager(this.rangeManager);
+  private manager: ColumnManager;
+  private supportView: SupportView;
+  private idtypes: IDType[];
+  private rowRange: Range1D;
+  private colRange: Range1D;
+  private newSupportView: SupportView;
+  private dataSize;
 
+  constructor(parent: HTMLElement) {
+    this.node = parent;
+  }
+
+  async build() {
+    await this.buildStartSelection(select('#startSelection'));
 
   }
 
-  /**
-   * Initialize the view and return a promise
-   * that is resolved as soon the view is completely initialized.
-   * @returns {Promise<App>}
-   */
-  init() {
+  private async buildStartSelection(elem: d3.Selection<any>) {
+    // get all idtypes, filter to the valid ones and SORT by name
+    const data: IDType[] = (await listAll())
+      .filter((d) => d instanceof IDType)
+      .map((d) => <IDType>d)
+      .sort((a, b) => a.name.localeCompare(b.name));
 
-    return this.build();
-  }
-
-  /**
-   * Load and initialize all necessary views
-   * @returns {Promise<App>}
-   */
-  private build() {
-    this.setBusy(true);
-    const blockList = new Map();
-    this.$node.select('main').append('div').classed('visManager', true);
-    this.visManager.filterManager = this.filterManager;
-
-
-    return listData().then((datasets) => {
-      datasets = convertTableToVectors(datasets);
-      console.log(datasets);
-      this.$node.select('h3').remove();
-      this.$node.select('button.adder').on('click', () => {
-        choose(datasets.map((d) => d.desc.name), 'Choose dataset').then((selection) => {
-          this.addDataset(datasets.find((d) => d.desc.name === selection));
-        });
+    this.idtypes = data;
+    // d3 binding to the dialog
+    const elems = elem.select('div.btn-group[role="group"]').selectAll('div.btn-group').data(data);
+    elems.enter().append('div')
+      .classed('btn-group', true)
+      .attr('role', 'group')
+      .html(`<button type="button" class="btn btn-default btn-lg">Artists</button>`);
+    elems.select('button')
+      .text((d) => d.names)
+      .on('click', (d) => {
+        this.setPrimaryIDType(d);
       });
-      this.setBusy(false);
+    elems.exit().remove();
+  }
+
+  private hideSelection() {//remove start selection
+    const elem = <HTMLElement>this.node.querySelector('#startSelection');
+    elem.style.display = 'none';
+  }
+
+  private showSelection() {
+    const elem = <HTMLElement>this.node.querySelector('#startSelection');
+    elem.style.display = null;
+  }
+
+  private reset() {
+    this.supportView.destroy();
+    this.manager.destroy();
+    this.removePreviewData();
+    this.showSelection();
+  }
+
+
+  private findType(data: IMotherTableType, currentIDType: string) {
+    const coltype = data.desc.coltype;
+    const rowtype = data.desc.rowtype;
+    if (rowtype === currentIDType) {
+
+      const idType = this.idtypes.filter((d) => d.id === coltype);
+      return idType[0];
+
+    } else if (coltype === currentIDType) {
+      const idType = this.idtypes.filter((d) => d.id === rowtype);
+      return idType[0];
+    }
+  }
+
+  private primarySortCol(evt: any, sortColdata: IAnyVector) {
+    this.supportView.primarySortColumn(sortColdata);
+
+  }
+
+  private setPrimaryIDType(idtype: IDType) {
+    this.hideSelection();
+    // create a column manager
+    this.manager = new ColumnManager(idtype, EOrientation.Horizontal, <HTMLElement>this.node.querySelector('main'));
+    this.manager.on(AVectorColumn.EVENT_PRIMARY_SORT_COLUMN, this.primarySortCol.bind(this));
+
+
+    const newdiv = document.createElement('div');
+    newdiv.classList.add(`support-view-${idtype.id}`);
+    const idName = document.createElement('div');
+    idName.classList.add('idType');
+    idName.innerHTML = (idtype.id.toUpperCase());
+    newdiv.appendChild(idName);
+    const previewDataNode = document.createElement('div');
+    previewDataNode.classList.add(`dataPreview-${idtype.id}`);
+    newdiv.appendChild(previewDataNode);
+
+
+    this.node.querySelector('section.rightPanel').appendChild(newdiv);
+
+    d3.select(previewDataNode).style('display', 'flex').append('div').classed('totalData', true);
+    d3.select(previewDataNode).append('div').classed('filteredData', true);
+
+
+    this.supportView = new SupportView(idtype, <HTMLElement>this.node.querySelector(`.support-view-${idtype.id}`));
+
+    this.supportView.on(FilterManager.EVENT_SORT_DRAGGING, (evt: any, data: AnyColumn[]) => {
+      this.manager.updateSortHierarchy(data);
+    });
+    // add to the columns if we add a dataset
+    this.supportView.on(SupportView.EVENT_DATASET_ADDED, (evt: any, data: IMotherTableType) => {
+      if (this.dataSize === undefined) {
+        this.dataSize = {total: (<any>data).indices.size(), filtered: (<any>data).indices.size()};
+        this.previewData(this.dataSize, idtype.id);
+      }
+
+      this.manager.push(data);
+      const checkMatrixType = data.desc.type;
+      if (checkMatrixType === 'matrix' && this.newSupportView === undefined || this.newSupportView === null) {
+        const otherIdtype: IDType = this.findType(data, idtype.id);
+        this.triggerMatrix();
+        this.newSupportManger(otherIdtype);
+      }
+
+    });
+
+
+    this.supportView.on(SupportView.EVENT_FILTER_CHANGED, (evt: any, filter: Range1D) => {
+      this.manager.filterData(filter);
+      // this.manager.update(filter);
+
+      this.rowRange = filter;
+      this.triggerMatrix();
+      this.dataSize.filtered = filter.size();
+      this.previewData(this.dataSize, idtype.id);
+
+    });
+
+    this.manager.on(ColumnManager.EVENT_DATA_REMOVED, (evt: any, data: IMotherTableType) => {
+      const cols = this.manager.columns;
+      const countSame = cols.filter((d, i) => d.data.desc.id === data.desc.id).length;
+      if (countSame < 1) {
+        this.supportView.remove(data);
+      }
+
+      if (this.manager.length === 0) {
+        this.reset();
+      }
     });
   }
 
 
-  private addDataset(data: IDataType) {
+  private newSupportManger(otherIdtype: IDType) {
 
-    const id = randomId();
+    // this.newManager = new ColumnManager(otherIdtype, EOrientation.Horizontal, <HTMLElement>this.node.querySelector('main'));
 
-    const currentRange: any = Block.currentRange;
+    const newdiv = document.createElement('div');
+    newdiv.classList.add(`support-view-${otherIdtype.id}`);
+    const idName = document.createElement('div');
+    idName.classList.add('idType');
+    idName.innerHTML = (otherIdtype.id.toUpperCase());
+    newdiv.appendChild(idName);
+    const previewDataNode = document.createElement('div');
+    previewDataNode.classList.add(`dataPreview-${otherIdtype.id}`);
+    newdiv.appendChild(previewDataNode);
+    this.node.querySelector('section.rightPanel').appendChild(newdiv);
 
-    <any>data.ids(currentRange).then((d) => {
+    d3.select(previewDataNode).style('display', 'flex').append('div').classed('totalData', true);
+    d3.select(previewDataNode).append('div').classed('filteredData', true);
 
-      (<any>data).idView(d).then((e) => {
+    this.newSupportView = new SupportView(otherIdtype, <HTMLElement>document.querySelector(`.support-view-${otherIdtype.id}`));
+    const m = this.supportView.matrixData;
+    const node = d3.select(`.${otherIdtype.id}.filter-manager`).append('div').classed('filter', true);
+    new MatrixFilter(m.t, <HTMLElement>node.node());
 
-        this.visManager.createVis(data, e, id);
-        this.filterManager.createFilter(App.blockList.get(id), this.filterManager);
+    this.previewData(this.dataSize, otherIdtype.id);
+    this.newSupportView.on(SupportView.EVENT_FILTER_CHANGED, (evt: any, filter: Range1D) => {
+      this.colRange = filter;
+      this.triggerMatrix();
 
-
-      });
-
+      this.dataSize.filtered = filter.size();
+      this.previewData(this.dataSize, otherIdtype.id);
     });
-
-    //this.visManager.createVis(data, data, id);  //first is new data and second is for filtered data purporse which is same as data at first
-
-    const filterNode = d3.select('#filterView');
-    // this.filterManager.createFilter(App.blockList.get(id), this.filterManager);
-    console.log(App.blockList);
-    //
-
-
-    // // ((<any>sorta).data().then((d) => console.log(d)))
 
 
   }
 
+  private triggerMatrix() {
+    const matrixCol = this.manager.columns.filter((d) => d instanceof MatrixColumn);
+    if (matrixCol.length === 0) {
+      return;
+    }
+    const indices = (<any>matrixCol[0]).data.indices;
+    if (this.rowRange === undefined) {
 
-  /**
-   * Show or hide the application loading indicator
-   * @param isBusy
-   */
-  setBusy(isBusy) {
-    this.$node.select('.busy').classed('hidden', !isBusy);
+      this.rowRange = (indices.dim(0));
+
+    }
+
+    if (this.colRange === undefined) {
+      this.colRange = (indices.dim(1));
+
+    }
+    // console.log(this.colRange)
+
+    matrixCol.forEach((col: MatrixColumn) => {
+      col.updateRows(this.rowRange);
+      col.updateCols(this.colRange);
+      col.updateMatrix(this.rowRange, this.colRange);
+    });
+
+  }
+
+  private previewData(dataSize, idtype) {
+    const availableWidth = parseFloat(d3.select(`.dataPreview-${idtype}`).style('width'));
+    const total = (dataSize.total)[0];
+    const filtered = (dataSize.filtered)[0] || 0;
+    const totalWidth = availableWidth / total * filtered;
+    const d = d3.select(`.dataPreview-${idtype}`);
+    d.style('height', '10px');
+    d3.select(`.dataPreview-${idtype}`).select('.totalData').style('width', `${totalWidth}px`);
+    d3.select(`.dataPreview-${idtype}`).select('.filteredData').style('width', `${availableWidth - totalWidth}px`);
+
+
+  }
+
+  private  removePreviewData() {
+    d3.selectAll('.rightPanel').remove();
+
   }
 
 }
@@ -123,6 +258,6 @@ export default class App {
  * @param parent
  * @returns {App}
  */
-export function create(parent: Element) {
+export function create(parent: HTMLElement) {
   return new App(parent);
 }
