+++ src/deprecated/RangeManager.ts
