+++ src/index.ts
@@ -8,13 +8,13 @@ import 'file-loader?name=robots.txt!./robots.txt';
 import 'phovea_ui/src/_bootstrap';
 import 'phovea_ui/src/_font-awesome';
 import './style.scss';
-import {create as createApp} from './app2';
+import {create as createApp} from './app';
 import {create as createHeader, AppHeaderLink} from 'phovea_ui/src/header';
 import {APP_NAME} from './language';
 
 const header = createHeader(
   <HTMLElement>document.querySelector('#caleydoHeader'),
-  { appLink: new AppHeaderLink(APP_NAME) }
+  {appLink: new AppHeaderLink(APP_NAME)}
 );
 
 const parent = <HTMLElement>document.querySelector('#app');
